Corrected weights for rate by system
Corrected the weights for ILI rate by system, so that they reflect the proportion of rural/urban population in every NUTS1 area in the total rural/urban population, rather than in the total population.
Which means that now the ILI rates by system are "on the same scale", and not additive to the total (total is a weighted mean of them).